mactry/last_try_OCR/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "last-try-ocr"
version = "2.0.0"
description = "Government-grade hybrid OCR for Bangla + English PDFs"
requires-python = ">=3.11"
dependencies = [
    "PyMuPDF>=1.23.0",
    "pdfplumber>=0.11.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "easyocr>=1.7.0",
    "google-genai>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[project.scripts]
last-try-ocr = "last_try_OCR.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"
